5-1_projects/02_ec_products_to_sheets/requirements.txt
python-dotenv
requests
beautifulsoup4
lxml
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pytz
